[Feat][#16] 예외 공통로직 정의
Contents
- 공통로직 처리 정의

Todo
- 예외 정의
