+++ backend/src/main/java/com/example/backend/controller/card/dto/DoingItem.java
@@ -1,8 +1,8 @@
-package com.example.backend.controller.card;
+package com.example.backend.controller.card.dto;
 
 import java.time.LocalDateTime;
 
-public class HaveDoneItem {
+public class DoingItem {
     private Long id;
     private String title;
     private String content;
@@ -10,7 +10,7 @@ public class HaveDoneItem {
     private LocalDateTime createdAt;
     private LocalDateTime lastModifiedAt;
 
-    public HaveDoneItem(Long id, String title, String content, String cardType, LocalDateTime createdAt, LocalDateTime lastModifiedAt) {
+    public DoingItem(Long id, String title, String content, String cardType, LocalDateTime createdAt, LocalDateTime lastModifiedAt) {
         this.id = id;
         this.title = title;
         this.content = content;
@@ -23,47 +23,23 @@ public class HaveDoneItem {
         return id;
     }
 
-    public void setId(Long id) {
-        this.id = id;
-    }
-
     public String getTitle() {
         return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
     }
 
     public String getContent() {
         return content;
     }
 
-    public void setContent(String content) {
-        this.content = content;
-    }
-
     public String getCardType() {
         return cardType;
-    }
-
-    public void setCardType(String cardType) {
-        this.cardType = cardType;
     }
 
     public LocalDateTime getCreatedAt() {
         return createdAt;
     }
 
-    public void setCreatedAt(LocalDateTime createdAt) {
-        this.createdAt = createdAt;
-    }
-
     public LocalDateTime getLastModifiedAt() {
         return lastModifiedAt;
-    }
-
-    public void setLastModifiedAt(LocalDateTime lastModifiedAt) {
-        this.lastModifiedAt = lastModifiedAt;
     }
 }
