+++ backend/src/main/java/com/example/backend/controller/exception/BaseExceptionType.java
@@ -1,0 +1,11 @@
+package com.example.backend.controller.exception;
+
+import org.springframework.http.HttpStatus;
+
+public interface BaseExceptionType {
+    String getErrorCode();
+
+    String getMessage();
+
+    HttpStatus getHttpStatus();
+}
