+++ backend/src/main/java/com/example/backend/controller/exception/BusinessException.java
@@ -1,0 +1,15 @@
+package com.example.backend.controller.exception;
+
+public class BusinessException extends RuntimeException {
+
+    private final BaseExceptionType baseExceptionType;
+
+    public BusinessException(BaseExceptionType baseExceptionType) {
+        super(baseExceptionType.getMessage());
+        this.baseExceptionType = baseExceptionType;
+    }
+
+    public BaseExceptionType getBaseExceptionType() {
+        return baseExceptionType;
+    }
+}
