+++ backend/src/main/java/com/example/backend/repository/card/jdbc/CardReadRepository.java
@@ -1,11 +1,13 @@
-package com.example.backend.controller.card;
+package com.example.backend.repository.card.jdbc;
 
+import com.example.backend.controller.card.dto.DoingItem;
+import com.example.backend.controller.card.dto.HaveDoneItem;
+import com.example.backend.controller.card.dto.TodoItem;
 import com.example.backend.domain.card.CardType;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 import static com.example.backend.utils.TimeUtils.dateTimeOf;
